[HMC5883L] compass sensor Now is posible get measurement of magnetic earth field, it is necessary configure the acceleration and gyro with the next register:
* Register for bypass mode:
  0x37 : 0x02
  0x64 : 0x00

*Register for disable sleep mode
 0x6B : 0x00

+++ kit-stm32f4/Standard_Library/STM32F4_ST7796-Display/configSPI.c
@@ -143,7 +143,7 @@ void SPI_writeData(uint8_t reg, uint8_t *Datavalue, uint32_t length){
 	}
 }
 
-void SPI_writePixel(uint8_t *Datavalue, uint32_t length){
+void SPI_writePixel(uint16_t *Datavalue, uint32_t length){
 
 	if (length <= 1){
 	  return;
