+++ kit-stm32f4/Standard_Library/STM32F4_ST7796-Display/configSPI.h
@@ -11,6 +11,6 @@ void SPI_writeReg(uint8_t reg, uint8_t regvalue);
 uint8_t SPI_readReg(uint8_t reg);
 void SPI_TxData(uint8_t value);
 void SPI_writeData(uint8_t reg, uint8_t *Datavalue, uint32_t length);
-void SPI_writePixel(uint8_t *Datavalue, uint32_t length);
+void SPI_writePixel(uint16_t *Datavalue, uint32_t length);
 
 #endif
